# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise AI Workspace",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Listing",
        "operationId": "listing_api_v1_workspaces__workspace_id__agents_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowContract"
                  },
                  "title": "Response Listing Api V1 Workspaces  Workspace Id  Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_workspaces__workspace_id__agents_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowContract"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/agents/{agent_id}/versions": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Revise",
        "operationId": "revise_api_v1_workspaces__workspace_id__agents__agent_id__versions_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowContract"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/agents/{agent_id}/duplicate": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Duplicate",
        "operationId": "duplicate_api_v1_workspaces__workspace_id__agents__agent_id__duplicate_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowContract"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/agents/{agent_id}/submit": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Submit",
        "operationId": "submit_api_v1_workspaces__workspace_id__agents__agent_id__submit_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowContract"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/agents/{agent_id}": {
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Remove",
        "operationId": "remove_api_v1_workspaces__workspace_id__agents__agent_id__delete",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "knowledge_document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Knowledge Document Ids"
          },
          "instructions": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Instructions"
          },
          "starters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Starters"
          },
          "tools": {
            "items": {
              "type": "string",
              "enum": [
                "knowledge_search",
                "read_source",
                "web_search",
                "web_read",
                "gmail_search",
                "gmail_read",
                "drive_search",
                "drive_read"
              ]
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "instructions"
        ],
        "title": "AgentInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeConfig": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 8000,
            "title": "Prompt",
            "default": "Answer the question using only the supplied evidence. Cite your sources."
          },
          "knowledge_document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Knowledge Document Ids"
          },
          "top_k": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "max_tool_calls": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Max Tool Calls",
            "default": 4
          },
          "condition": {
            "type": "string",
            "enum": [
              "has_sources",
              "has_answer",
              "abstained"
            ],
            "title": "Condition",
            "default": "has_sources"
          },
          "tools": {
            "items": {
              "type": "string",
              "enum": [
                "knowledge_search",
                "read_source",
                "web_search",
                "web_read",
                "gmail_search",
                "gmail_read",
                "drive_search",
                "drive_read"
              ]
            },
            "type": "array",
            "maxItems": 8,
            "title": "Tools"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "NodeConfig"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowContract": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused"
          },
          "published_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Version Id"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/WorkflowVersionContract"
            },
            "type": "array",
            "title": "Versions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "workflow"
          },
          "starters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Starters"
          },
          "submitted_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted From"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "name",
          "description",
          "paused",
          "published_version_id",
          "versions",
          "created_at"
        ],
        "title": "WorkflowContract"
      },
      "WorkflowDefinition": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNode"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 2,
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEdge"
            },
            "type": "array",
            "maxItems": 60,
            "minItems": 1,
            "title": "Edges"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "WorkflowDefinition"
      },
      "WorkflowEdge": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "true",
                  "false"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source",
          "target"
        ],
        "title": "WorkflowEdge"
      },
...
      "WorkflowNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z][a-zA-Z0-9_-]{0,63}$",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "input",
              "retrieval",
              "model",
              "condition",
              "agent",
              "answer"
            ],
            "title": "Type"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label"
          },
          "position": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Position"
          },
          "config": {
            "$ref": "#/components/schemas/NodeConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "label"
        ],
        "title": "WorkflowNode"
      },
      "WorkflowVersionContract": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "definition": {
            "$ref": "#/components/schemas/WorkflowDefinition"
          },
          "corpus_revision": {
            "type": "integer",
            "title": "Corpus Revision"
          },
          "model_profile": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "object",
            "title": "Model Profile"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workflow_id",
          "number",
          "definition",
          "corpus_revision",
          "model_profile",
          "fingerprint",
          "created_at",
          "created_by"
        ],
        "title": "WorkflowVersionContract"
      },
...
    }
  }
}